+++ common/src/main/java/org/conductoross/conductor/common/webhook/model/IncomingWebhookEvent.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.conductoross.conductor.common.webhook;
+package org.conductoross.conductor.common.webhook.model;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
