+++ core/src/main/java/org/conductoross/conductor/dao/webhook/WebhookConfigDAO.java
@@ -10,10 +10,13 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.conductoross.conductor.common.webhook;
+package org.conductoross.conductor.dao.webhook;
 
 import java.util.List;
 import java.util.Map;
+
+import org.conductoross.conductor.common.webhook.model.IncomingWebhookEvent;
+import org.conductoross.conductor.common.webhook.model.WebhookConfig;
 
 /**
  * Persists webhook configurations, matcher indexes, and inbound event audit records.
