+++ webhook-task/src/main/java/org/conductoross/conductor/tasks/webhook/verifier/WebhookVerifier.java
@@ -10,11 +10,14 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.conductoross.conductor.common.webhook;
+package org.conductoross.conductor.tasks.webhook.verifier;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+
+import org.conductoross.conductor.common.webhook.model.IncomingWebhookEvent;
+import org.conductoross.conductor.common.webhook.model.WebhookConfig;
 
 /**
  * Verifies inbound webhook requests against a {@link WebhookConfig}.
